src/CompanionCube.LlmBridge/Python/requirements.txt
flask==2.3.3
llama-cpp-python==0.2.11
numpy==1.24.3
werkzeug==2.3.7